+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.11.0
+* fix error in the javaFuncs helper
+* extend internal model with minItems and maxItems restrictions for array properties
+* extend internal model with uniqueItems for array properties
+
+# 0.10.0
+* improvements for the Java bean template
+
 # 0.9.0
 * include implicit references for properties